fast_apis/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==18.2.0
axios==0.4.0
azure-core==1.35.0
azure-identity==1.24.0
backports.tarfile==1.2.0
banks==2.2.0
beautifulsoup4==4.13.4
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpickle==3.1.1
colabgymrender==1.0.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
crashtest==0.4.1
cryptography==45.0.6
cvxopt==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.3.2
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.4
griffe==1.11.1
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imbalanced-learn==0.13.0
immutabledict==4.2.1
importlib_metadata==8.7.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.35.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
keyring==25.6.0
kiwisolver==1.4.8
libclang==18.1.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.1
llama-index-agent-openai==0.4.8
llama-index-cli==0.5.0
llama-index-core==0.13.1
llama-index-embeddings-azure-openai==0.3.2
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.1
llama-index-instrumentation==0.4.0
llama-index-llms-azure-openai==0.3.2
llama-index-llms-openai==0.5.3
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.5.1
llama-index-readers-llama-parse==0.5.0
llama-index-workflows==1.3.0
llama-parse==0.6.21
lxml==6.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
MouseInfo==0.1.3
moviepy==2.2.1
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
numpy==1.23.5
openai==1.99.9
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
ortools==9.12.4544
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pbs-installer==2025.4.9
pdfkit==1.0.0
pdflatex==0.1.3
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
proglog==0.1.12
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==3.0.4
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pypdf==5.9.0
pyperclip==1.9.0
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.1
PyVirtualDisplay==3.0
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
regex==2024.11.6
reportlab==4.4.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sklearn-crfsuite==0.5.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
threadpoolctl==3.5.0
tiktoken==0.11.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
torchviz==0.0.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
trove-classifiers==2025.4.28.22
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
wcwidth==0.2.13
Werkzeug==3.1.3
wkhtmltopdf==0.2
wrapt==1.17.2
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
